+++ app/src/main/java/remix/myplayer/db/room/model/PlayList.kt
@@ -1,6 +1,6 @@
 package remix.myplayer.db.room.model
 
-import android.arch.persistence.room.*
+import androidx.room.*
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 
@@ -11,7 +11,7 @@ import com.google.gson.reflect.TypeToken
 @TypeConverters(PlayList.Converter::class)
 data class PlayList(
     @PrimaryKey(autoGenerate = true)
-    val id: Int,
+    val id: Long,
     val name: String,
 //    val count: Int,
     val audioIds: LinkedHashSet<Int>,
