+++ app/src/main/java/remix/myplayer/ui/activity/CustomSortActivity.kt
@@ -3,11 +3,11 @@ package remix.myplayer.ui.activity
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.support.design.widget.FloatingActionButton
-import android.support.v7.widget.DefaultItemAnimator
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
-import android.support.v7.widget.helper.ItemTouchHelper
+import com.google.android.material.floatingactionbutton.FloatingActionButton
+import androidx.recyclerview.widget.DefaultItemAnimator
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.ItemTouchHelper
 import android.view.View
 import butterknife.ButterKnife
 import butterknife.OnClick
@@ -43,14 +43,14 @@ class CustomSortActivity : ToolbarActivity() {
   }
 
   private lateinit var songs: ArrayList<Song>
-  private var playlistId = -1
+  private var playlistId = -1L
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     setContentView(R.layout.activity_custom_sort)
     ButterKnife.bind(this@CustomSortActivity)
 
-    playlistId = intent.getIntExtra(EXTRA_ID, -1)
+    playlistId = intent.getLongExtra(EXTRA_ID, -1)
     val playlistName = intent.getStringExtra(EXTRA_NAME) ?: ""
 
     if (!intent.hasExtra(EXTRA_LIST)) {
@@ -134,7 +134,7 @@ class CustomSortActivity : ToolbarActivity() {
 
   companion object {
     @JvmStatic
-    fun start(context: Context, id: Int, name: String, list: List<Song>) {
+    fun start(context: Context, id: Long, name: String, list: List<Song>) {
       context.startActivity(Intent(context, CustomSortActivity::class.java)
           .putExtra(EXTRA_ID, id)
           .putExtra(EXTRA_NAME, name)
