+++ app/src/main/java/remix/myplayer/ui/adapter/AlbumAdapter.java
@@ -5,10 +5,10 @@ import static remix.myplayer.request.ImageUriRequest.SMALL_IMAGE_SIZE;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.PopupMenu;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.widget.PopupMenu;
+import androidx.recyclerview.widget.RecyclerView;
 import android.text.TextUtils;
 import android.view.Gravity;
 import android.view.LayoutInflater;
@@ -87,7 +87,7 @@ public class AlbumAdapter extends HeaderAdapter<Album, BaseViewHolder> implement
     holder.mText1.setText(album.getAlbum());
 
     //设置封面
-    final int albumId = album.getAlbumID();
+    final long albumId = album.getAlbumID();
     final int imageSize = mMode == LIST_MODE ? SMALL_IMAGE_SIZE : BIG_IMAGE_SIZE;
     holder.mImage.setTag(setImage(holder.mImage, ImageUriUtil.getSearchRequest(album), imageSize, position));
